Extract MATCH responsibilities from MatchBuilder
