+++ arangod/Aql/MatchBuilder.h
@@ -22,13 +22,14 @@
 
 #pragma once
 
+#include "Aql/MatchCollectionAccessBuilder.h"
+#include "Aql/MatchFilterBuilder.h"
 #include "Aql/MatchPatternTypes.h"
+#include "Aql/MatchProjectionBuilder.h"
 #include "Aql/types.h"
 
 #include <cstddef>
 #include <optional>
-#include <span>
-#include <string_view>
 #include <tuple>
 #include <unordered_map>
 #include <vector>
@@ -40,12 +41,12 @@ struct AstNode;
 class CalculationNode;
 class ExecutionNode;
 class ExecutionPlan;
-class FilterNode;
 struct Variable;
 
-/// @brief Lowers normalized MATCH patterns into ExecutionPlan fragments.
-/// Parser AST semantics are normalized by MatchPatternNormalizer before
-/// planning; this class owns only execution-plan construction.
+/// @brief Orchestrates lowering of normalized MATCH patterns into
+/// ExecutionPlan fragments. Collection access, filtering, projection, and
+/// variable substitution are delegated to focused helpers; this class
+/// coordinates pattern construction and plan wiring.
 class MatchBuilder {
  public:
   MatchBuilder(ExecutionPlan& plan, Ast* ast);
@@ -54,49 +55,6 @@ class MatchBuilder {
   ExecutionNode* build(ExecutionNode* previous, AstNode const* matchNode);
 
  private:
-  AstNode* createPropertyAccess(Variable const* variable,
-                                std::string_view property);
-
-  AstNode* buildEdgeCollectionList(NormalizedEdge const& edge);
-
-  std::tuple<CalculationNode*, FilterNode*> createPropertiesFilter(
-      Variable const* variable,
-      std::vector<MatchPropertyConstraint> const& properties,
-      std::optional<MatchExpressionRef> const& additionalFilter,
-      std::unordered_map<VariableId, Variable const*> const& subst);
-
-  std::tuple<ExecutionNode*, ExecutionNode*, Variable const*>
-  createCollectionAccess(
-      NormalizedVertex const& vertex, Variable const* fullDocumentVariable,
-      std::unordered_map<VariableId, Variable const*> const& subst);
-
-  ExecutionNode* createDocumentPatternProjection(
-      Variable const* destinationVariable, Variable const* fullDocumentVar,
-      std::optional<MatchProjection> const& projection,
-      std::unordered_map<VariableId, Variable const*> const& subst);
-
-  ExecutionNode* createEdgeDocumentPatternProjection(
-      Variable const* destinationVariable, Variable const* fullDocumentVar,
-      std::optional<MatchProjection> const& projection,
-      std::unordered_map<VariableId, Variable const*> const& subst);
-
-  /// @brief Shared MATCH projection lowering. @p mandatoryAttributes are
-  /// auto-injected and treated as reserved for user projection handling.
-  ExecutionNode* createPatternProjection(
-      Variable const* destinationVariable, Variable const* fullDocumentVar,
-      std::optional<MatchProjection> const& projection,
-      std::span<std::string_view const> mandatoryAttributes,
-      std::unordered_map<VariableId, Variable const*> const& subst);
-
-  std::tuple<ExecutionNode*, ExecutionNode*, Variable const*>
-  createPatternEdgeEnumerateAccess(
-      NormalizedEdge const& edge, Variable const* outputVariable,
-      std::unordered_map<VariableId, Variable const*> const& subst);
-
-  std::tuple<CalculationNode*, FilterNode*> createVertexEdgeFilter(
-      Variable const* leftVertex, Variable const* edge,
-      Variable const* rightVertex, MatchEdgeDirection direction);
-
   /// @param edgeDocumentOutputVariable Edge document output for fixed-depth
   /// traversals. Ignored when the edge variable receives a path object.
   /// @param vertexDocumentOutputVariable Vertex output when @p target is a
@@ -131,6 +89,9 @@ class MatchBuilder {
 
   ExecutionPlan& _plan;
   Ast* _ast;
+  MatchFilterBuilder _filters;
+  MatchCollectionAccessBuilder _collections;
+  MatchProjectionBuilder _projections;
 };
 
 }  // namespace arangodb::aql
